refactor :  - auth 관련 파일 패키지 이동.  - user 엔티티 직렬화 추가. feat :  - 파라미터 레벨의 SessionUser 클래스에 @LoginUser 애너테이션을 붙이면 세션에 저장된 사용자 정보를 가져올 수 있도록 추가.
