+++ src/main/java/com/ajou/travely/config/auth/SecurityConfig.java
@@ -1,4 +1,4 @@
-package com.ajou.travely.config;
+package com.ajou.travely.config.auth;
 
 import com.ajou.travely.domain.user.Type;
 import lombok.RequiredArgsConstructor;
@@ -16,7 +16,8 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
         http.csrf().disable();
         http.authorizeRequests()
                 .antMatchers("/api/v1/oauth2/authorization/kakao", "/api/v1/isLogin", "/api/v1/signup").permitAll()
-                .antMatchers("/api/v1/**").access("hasRole('ROLE_USER') or hasRole('ROLE_ADMIN')");
+                .antMatchers("/api/v1/**").permitAll();
+//                .access("hasRole('ROLE_USER') or hasRole('ROLE_ADMIN')");
 //                .antMatchers("/api/v2/**").access("hasRole('ROLE_ADMIN')")
     }
 }
