+++ src/main/java/com/ajou/travely/config/auth/SessionUser.java
@@ -1,0 +1,12 @@
+package com.ajou.travely.config.auth;
+
+import lombok.Getter;
+
+@Getter
+public class SessionUser {
+    private Long userId;
+
+    public SessionUser(Long userId) {
+        this.userId = userId;
+    }
+}
