+++ src/main/java/com/ajou/travely/domain/user/User.java
@@ -1,5 +1,6 @@
 package com.ajou.travely.domain.user;
 
+import java.io.Serializable;
 import java.time.LocalDate;
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -16,7 +17,7 @@ import lombok.*;
 @NoArgsConstructor
 @Table(name = "users")
 @Entity
-public class User {
+public class User implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "user_id")
